rbnf can't handle small numbers
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/1722



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@14738 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/nfrule.cpp
@@ -321,7 +321,7 @@ NFRule::parseRuleDescriptor(UnicodeString& description, UErrorCode& status)
 
                 // tempValue now contain's the rule's radix.  Set it
                 // accordingly, and recalculate the rule's exponent
-                radix = (int16_t)val;
+                radix = (int32_t)val;
                 if (radix == 0) {
                     // throw new IllegalArgumentException("Rule can't have radix of 0");
                     status = U_PARSE_ERROR;
